use world.getTopSolidOrLiquidBlock to find top block
The use of chunk.getTopFilledSegment was incorrect; it sometimes returns
a height *below* the topmost block, which prevented spawning a reef
there. This was only found upon changing the code to make all topmost
blocks into reef, and discovering just how much of the world did not
convert. It was particularly bad around slopes and beaches.

Contrary to its name, getTopSolidOrLiquidBlock walks through all blocks
that don't block movement, plus leaves and foliage. We duplicate the
functionality in getTop on purpose, to allow spawning reefs under other
blocks in future (e.g. ice).